+++ src/main/java/com/BE/starTroad/service/JpaStudierService.java
@@ -18,14 +18,14 @@ public class JpaStudierService {
 
 
     //스터디 아이디로 몇 명이 스터디에 참여하는지 조회
-    public int studyNum(Long studyId) {
+    public int studyNum(int studyId) {
         List<Studier> studiers = new ArrayList<>();
-        springDataJpaStudierRepository.findByStudyId(studyId.intValue()).forEach(e -> studiers.add(e));
+        springDataJpaStudierRepository.findByStudyId(studyId).forEach(e -> studiers.add(e));
         int num = studiers.size();
         return num;
     }
 
-    public Optional<Studier> findAmIJoin(Long studyId, String email) {
+    public Optional<Studier> findAmIJoin(int studyId, String email) {
         Optional<Studier> studier = springDataJpaStudierRepository.findByStudyIdAndEmail(studyId, email);
         return studier;
     }
